pull

+++ src/pages/ProfilePage/MyOrderPage/OrderDetailModal.jsx
@@ -1,4 +1,4 @@
-import React, {useEffect, useState} from 'react';
+import React, {useEffect, useRef, useState} from 'react';
 
 import {
 	DeleteOutlined,
@@ -47,13 +47,16 @@ import {OrderLog} from './OrderLog';
 import {OrderPayment} from './OrderPayment';
 import InformationUser from './InformationUser';
 import {getTransactionByOrderId} from '../../../redux/slices/transactionSlice';
-import {useParams} from 'react-router-dom';
+import {useLocation, useNavigate, useParams} from 'react-router-dom';
 
 const {Text, Title} = Typography;
 
 export const OrderDetailModal = () => {
 	const {id} = useParams();
+	const location = useLocation();
+	const transactionRef = useRef(null);
 	const dispatch = useDispatch();
+	const navigate = useNavigate();
 	const loading = useSelector(LoadingOrderSelector);
 	const loadingReview = useSelector(LoadingReviewSelector);
 	const reviewDetail = useSelector(ReviewSelector);
@@ -87,6 +90,9 @@ export const OrderDetailModal = () => {
 		paymentMethod: order?.PaymentMethod?.MappedName,
 		shippingAddress: order?.ShippingAddress,
 		status: order?.Status,
+		warrantyPrice: formatPrice(item?.WarrantyPrice || 0),
+		discountSavedAmount: formatPrice(item?.DiscountSavedAmount || 0),
+		promotionSavedAmount: formatPrice(item?.PromotionSavedAmount || 0),
 		shippingFee: formatPrice(order?.ShippingFee || 0),
 		totalPrice: formatPrice(order?.TotalPrice || 0),
 		totalRefund: formatPrice(order?.TotalRefund || 0),
@@ -117,6 +123,39 @@ export const OrderDetailModal = () => {
 			render: (text) => <div className="flex justify-center">{text}</div>,
 		},
 		{
+			title: 'Giá Bảo Hành',
+			dataIndex: 'warrantyPrice',
+			key: 'warrantyPrice',
+			align: 'center',
+			render: (_, record) => (
+				<div className="flex flex-col items-center">
+					<div>{record.warrantyPrice}</div>
+				</div>
+			),
+		},
+		{
+			title: 'Giá Giảm',
+			dataIndex: 'discountSavedAmount',
+			key: 'discountSavedAmount',
+			align: 'center',
+			render: (_, record) => (
+				<div className="flex flex-col items-center">
+					<div>{record.discountSavedAmount}</div>
+				</div>
+			),
+		},
+		{
+			title: 'Giá Khuyến Mãi',
+			dataIndex: 'promotionSavedAmount',
+			key: 'promotionSavedAmount',
+			align: 'center',
+			render: (_, record) => (
+				<div className="flex flex-col items-center">
+					<div>{record.promotionSavedAmount}</div>
+				</div>
+			),
+		},
+		{
 			title: 'Giá Sản Phẩm',
 			dataIndex: 'price',
 			key: 'price',
@@ -128,7 +167,7 @@ export const OrderDetailModal = () => {
 			),
 		},
 		{
-			title: 'HT Giao Hàng',
+			title: 'Hình Thức Giao Hàng',
 			dataIndex: 'IsCollectAtShop',
 			key: 'IsCollectAtShop',
 			align: 'center',
@@ -245,6 +284,15 @@ export const OrderDetailModal = () => {
 	}, [dispatch, statusOrder, reviewDetail, transfer, cancelled]);
 
 	useEffect(() => {
+		if (location.state?.scrollTo) {
+			const targetSection = location.state.scrollTo;
+			if (targetSection === 'transaction') {
+				transactionRef.current.scrollIntoView({behavior: 'smooth'});
+			}
+		}
+	}, [location]);
+
+	useEffect(() => {
 		if (id) {
 			dispatch(getTransactionByOrderId(id))
 				.unwrap()
@@ -266,17 +314,27 @@ export const OrderDetailModal = () => {
 
 	const handleDeleteReview = () => {
 		if (reviewContent) {
-			dispatch(deleteReviewAction(reviewContent.Id))
-				.unwrap()
-				.then((res) => {
-					setReviewContent(null);
-					setRating(0);
-					message.success('Đánh giá đã được xóa');
-					setIsModalVisible(false);
-				})
-				.catch((error) => {
-					message.error(error?.detail || error?.data?.detail);
-				});
+			Modal.confirm({
+				title: 'Xác nhận xóa đánh giá',
+				content: 'Bạn có chắc chắn muốn xóa đánh giá này không?',
+				okText: 'Xóa',
+				centered: true,
+				cancelText: 'Hủy',
+				onOk: () => {
+					dispatch(deleteReviewAction(reviewContent.Id))
+						.unwrap()
+						.then((res) => {
+							setReviewContent(null);
+							setRating(0);
+							message.success('Đánh giá đã được xóa');
+							setIsModalVisible(false);
+						})
+						.catch((error) => {
+							message.error(error?.data?.detail);
+						});
+				},
+				onCancel: () => {},
+			});
 		}
 	};
 
@@ -287,6 +345,10 @@ export const OrderDetailModal = () => {
 	const handleReviewRequest = (id) => {
 		setJewelryId(id);
 		setIsReviewModalVisible(true);
+	};
+
+	const handleFileRemove = (file) => {
+		setImageFiles((prev) => prev.filter((item) => item.uid !== file.uid));
 	};
 
 	const submitCancelOrder = (values) => {
@@ -382,7 +444,7 @@ export const OrderDetailModal = () => {
 							<Title level={3} className="mb-4">
 								Thông Tin Khách Hàng
 							</Title>
-							<InformationUser order={order} />
+							<InformationUser order={order} transaction={transaction} />
 						</div>
 						{order?.Deliverer && (
 							<div className="my-5">
@@ -406,7 +468,7 @@ export const OrderDetailModal = () => {
 						)}
 
 						<div className="w-full flex flex-col sm:flex-row gap-4">
-							<div className="w-full sm:w-2/3">
+							<div ref={transactionRef} className="w-full sm:w-2/3">
 								{order?.Status === 1 && order?.Transactions?.length === 0 ? (
 									<OrderPayment order={order} />
 								) : (
@@ -482,7 +544,15 @@ export const OrderDetailModal = () => {
 								Tổng cộng: {formatPrice(order?.TotalPrice)}
 							</div>
 						</div>
-						<Button type="text" className="text-primary" icon={<LeftOutlined />}>
+						<Button
+							onClick={() => {
+								navigate('/my-orders');
+								localStorage.setItem('lastVisitedPage', 'Đơn hàng của tôi');
+							}}
+							type="text"
+							className="text-primary"
+							icon={<LeftOutlined />}
+						>
 							Quay lại danh sách đơn hàng của tôi
 						</Button>
 						{/* Cancel Order Modal */}
@@ -538,8 +608,9 @@ export const OrderDetailModal = () => {
 										listType="picture"
 										fileList={fileList}
 										onChange={handleFileChange}
+										onRemove={handleFileRemove}
 										beforeUpload={(file) => {
-											setImageFiles((fileList) => [...fileList, file]);
+											setImageFiles((prev) => [...prev, file]);
 											return false;
 										}}
 										maxCount={3}
@@ -579,6 +650,7 @@ export const OrderDetailModal = () => {
 				title="Thông tin Đánh Giá"
 				visible={isModalVisible}
 				onCancel={handleCancel}
+				centered
 				footer={[
 					<Button
 						key="delete"
